Add comprehensive scikit-learn models with OpenVINO optimization and update documentation
