+++ src/otx/backend/openvino/models/sklearn/classification/knn.py
@@ -1,42 +1,29 @@
 # Copyright (C) 2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-"""Class definition for the SVR regression model with OpenVINO optimization for OTX."""
+"""K-Nearest Neighbors Classifier model wrapper with OpenVINO optimization for OTX."""
 
-from sklearnex import patch_sklearn
-import os
 import joblib
 from time import time
-from sklearn.svm import SVR as SkModel
-from sklearn.metrics import r2_score
+from sklearnex.neighbors import KNeighborsClassifier as SkModel
+from sklearn.metrics import classification_report, accuracy_score
 
-class SVR:
+class KNeighborsClassifier:
     def __init__(self, *args, use_openvino=True, **kwargs):
         """
-        Initialize the SVR wrapper.
+        Initialize the KNeighborsClassifier wrapper.
 
         Args:
-            *args: Positional arguments for sklearn's SVR.
+            *args: Positional arguments for sklearn's KNeighborsClassifier.
             use_openvino (bool): Whether to enable OpenVINO optimizations.
-            **kwargs: Keyword arguments for sklearn's SVR.
+            **kwargs: Keyword arguments for sklearn's KNeighborsClassifier.
         """
         self.use_openvino = use_openvino
-        self._patched = False
-        self._ir_model = None
-
-        if self.use_openvino:
-            try:
-                patch_sklearn()
-                self._patched = True
-                print("✅ sklearnex patch applied successfully.")
-            except Exception:
-                print("⚠️ sklearnex patch failed.")
-
         self.model = SkModel(*args, **kwargs)
-        print("📦 SVR model initialized.")
+        print("📦 KNeighborsClassifier model initialized (sklearnex version).")
 
     def fit(self, X, y):
         """
-        Fit the SVR model.
+        Fit the KNeighborsClassifier model.
 
         Args:
             X (array-like): Training data.
@@ -49,49 +36,51 @@ class SVR:
 
     def predict(self, X):
         """
-        Predict target values for samples in X.
+        Predict class labels for samples in X.
 
         Args:
             X (array-like): Input data.
 
         Returns:
-            array: Predicted values.
+            array: Predicted class labels.
         """
         return self.model.predict(X)
 
     def score(self, X, y):
         """
-        Returns the R2 score on the given test data and labels.
+        Returns the mean accuracy on the given test data and labels.
 
         Args:
             X (array-like): Test samples.
-            y (array-like): True values for X.
+            y (array-like): True labels for X.
 
         Returns:
-            float: R2 score.
+            float: Mean accuracy.
         """
-        score = r2_score(y, self.predict(X))
-        print(f"📊 R2 Score: {score:.4f}")
-        return score
+        acc = self.model.score(X, y)
+        print(f"📊 Model score: {acc:.4f}")
+        return acc
 
     def evaluate(self, X, y):
         """
-        Evaluate the model and print inference time.
+        Evaluate the model and print inference time and classification report.
 
         Args:
             X (array-like): Test samples.
-            y (array-like): True values for X.
+            y (array-like): True labels for X.
 
         Returns:
-            float: R2 score.
+            float: Mean accuracy.
         """
         start = time()
-        score = self.score(X, y)
+        y_pred = self.predict(X)
         elapsed = time() - start
-        print(f"📈 Inference time: {elapsed:.4f} seconds.")
-        return score
+        acc = accuracy_score(y, y_pred)
+        print(f"📈 Accuracy: {acc:.4f} | Inference time: {elapsed:.4f} seconds.")
+        print(classification_report(y, y_pred))
+        return acc
 
-    def save_model(self, path="svr_model.joblib"):
+    def save_model(self, path="knn_model.joblib"):
         """
         Save the trained model to a file.
 
@@ -101,7 +90,7 @@ class SVR:
         joblib.dump(self.model, path)
         print(f"💾 Model saved to {path}")
 
-    def load_model(self, path="svr_model.joblib"):
+    def load_model(self, path="knn_model.joblib"):
         """
         Load a model from a file.
 
@@ -111,12 +100,12 @@ class SVR:
         self.model = joblib.load(path)
         print(f"📂 Model loaded from {path}")
 
-    def convert_to_ir(self, X_train, model_name="svr"):
+    def convert_to_ir(self, X_train, model_name="knn_model"):
         """
-        Not supported: Exporting SVR to IR via neural network is not possible.
+        Not supported: Exporting KNeighborsClassifier to IR via neural network is not possible.
 
         Args:
             X_train (array-like): Training data (unused).
             model_name (str): Model name (unused).
         """
-        print("❌ Export to IR via neural network is not supported for SVR.")
+        print("❌ Export to IR via neural network is not supported for KNeighborsClassifier.")
