+++ src/otx/backend/openvino/models/sklearn/classification/nusvc.py
@@ -1,40 +1,34 @@
 # Copyright (C) 2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-"""K-Nearest Neighbors Classifier model wrapper with OpenVINO optimization for OTX."""
+"""Class definition for the Nu-Support Vector Classifier with OpenVINO optimization for OTX."""
 
-from sklearnex import patch_sklearn
 import joblib
+import numpy as np
 from time import time
-from sklearn.neighbors import KNeighborsClassifier as SkModel
-from sklearn.metrics import classification_report, accuracy_score
+from sklearnex.svm import NuSVC as SkNuSVC
+from sklearn.metrics import accuracy_score
+import warnings
+from sklearn.exceptions import ConvergenceWarning
+warnings.filterwarnings("ignore", category=ConvergenceWarning)
 
-class KNeighborsClassifier:
+class NuSVC:
     def __init__(self, *args, use_openvino=True, **kwargs):
         """
-        Initialize the KNeighborsClassifier wrapper.
+        Initialize the NuSVC wrapper.
 
         Args:
-            *args: Positional arguments for sklearn's KNeighborsClassifier.
+            *args: Positional arguments for sklearn's NuSVC.
             use_openvino (bool): Whether to enable OpenVINO optimizations.
-            **kwargs: Keyword arguments for sklearn's KNeighborsClassifier.
+            **kwargs: Keyword arguments for sklearn's NuSVC.
         """
         self.use_openvino = use_openvino
-        self._patched = False
-
-        if self.use_openvino:
-            try:
-                patch_sklearn()
-                self._patched = True
-                print("✅ sklearnex patch applied successfully.")
-            except Exception:
-                print("⚠️ sklearnex patch failed.")
-
-        self.model = SkModel(*args, **kwargs)
-        print("📦 KNeighborsClassifier model initialized.")
+        self._ir_model = None
+        self.model = SkNuSVC(*args, **kwargs)
+        print("📦 NuSVC model initialized (sklearnex version).")
 
     def fit(self, X, y):
         """
-        Fit the KNeighborsClassifier model.
+        Fit the NuSVC model.
 
         Args:
             X (array-like): Training data.
@@ -57,6 +51,24 @@ class KNeighborsClassifier:
         """
         return self.model.predict(X)
 
+    def predict_proba(self, X):
+        """
+        Probability estimates for samples in X.
+
+        Args:
+            X (array-like): Input data.
+
+        Returns:
+            array: Probability estimates.
+
+        Raises:
+            AttributeError: If probability estimates are not supported.
+        """
+        if hasattr(self.model, "predict_proba"):
+            return self.model.predict_proba(X)
+        else:
+            raise AttributeError("This NuSVC model does not support probability estimates.")
+
     def score(self, X, y):
         """
         Returns the mean accuracy on the given test data and labels.
@@ -68,13 +80,13 @@ class KNeighborsClassifier:
         Returns:
             float: Mean accuracy.
         """
-        acc = self.model.score(X, y)
-        print(f"📊 Model score: {acc:.4f}")
+        acc = accuracy_score(y, self.predict(X))
+        print(f"📊 Accuracy: {acc:.4f}")
         return acc
 
     def evaluate(self, X, y):
         """
-        Evaluate the model and print inference time and classification report.
+        Evaluate the model and print inference time.
 
         Args:
             X (array-like): Test samples.
@@ -84,14 +96,12 @@ class KNeighborsClassifier:
             float: Mean accuracy.
         """
         start = time()
-        y_pred = self.predict(X)
+        acc = self.score(X, y)
         elapsed = time() - start
-        acc = accuracy_score(y, y_pred)
-        print(f"📈 Accuracy: {acc:.4f} | Inference time: {elapsed:.4f} seconds.")
-        print(classification_report(y, y_pred))
+        print(f"📈 Inference time: {elapsed:.4f} seconds.")
         return acc
 
-    def save_model(self, path="knn_model.joblib"):
+    def save_model(self, path="nusvc_model.joblib"):
         """
         Save the trained model to a file.
 
@@ -101,7 +111,7 @@ class KNeighborsClassifier:
         joblib.dump(self.model, path)
         print(f"💾 Model saved to {path}")
 
-    def load_model(self, path="knn_model.joblib"):
+    def load_model(self, path="nusvc_model.joblib"):
         """
         Load a model from a file.
 
@@ -111,12 +121,12 @@ class KNeighborsClassifier:
         self.model = joblib.load(path)
         print(f"📂 Model loaded from {path}")
 
-    def convert_to_ir(self, X_train, model_name="knn_model"):
+    def convert_to_ir(self, X_train, model_name="nusvc_model"):
         """
-        Not supported: Exporting KNeighborsClassifier to IR via neural network is not possible.
+        Not supported: Exporting NuSVC to IR via neural network is not possible.
 
         Args:
             X_train (array-like): Training data (unused).
             model_name (str): Model name (unused).
         """
-        print("❌ Export to IR via neural network is not supported for KNeighborsClassifier.")
+        print("❌ Export to IR via neural network is not supported for NuSVC.")
