+++ src/otx/backend/openvino/models/sklearn/decomposition/pca.py
@@ -1,40 +1,30 @@
 # Copyright (C) 2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-"""Class definition for the Incremental PCA model with OpenVINO optimization for OTX."""
+"""Class definition for the PCA model with OpenVINO optimization for OTX."""
 
-from sklearnex import patch_sklearn
 import joblib
 from time import time
-from sklearn.decomposition import IncrementalPCA as SkIncrementalPCA
+from sklearnex.decomposition import PCA as SkPCA
 
-class IncrementalPCA:
+class PCA:
     def __init__(self, *args, use_openvino=True, **kwargs):
         """
-        Initialize the IncrementalPCA wrapper.
+        Initialize the PCA wrapper.
 
         Args:
-            *args: Positional arguments for sklearn's IncrementalPCA.
+            *args: Positional arguments for sklearn's PCA.
             use_openvino (bool): Whether to enable OpenVINO optimizations.
-            **kwargs: Keyword arguments for sklearn's IncrementalPCA.
+            **kwargs: Keyword arguments for sklearn's PCA.
         """
         self.use_openvino = use_openvino
-        self._patched = False
         self._ir_model = None
 
-        if self.use_openvino:
-            try:
-                patch_sklearn()
-                self._patched = True
-                print("✅ sklearnex patch applied successfully.")
-            except ImportError:
-                print("⚠️ sklearnex is not installed. Running without OpenVINO optimization.")
-
-        self.model = SkIncrementalPCA(*args, **kwargs)
-        print("📦 IncrementalPCA model initialized.")
+        self.model = SkPCA(*args, **kwargs)
+        print("📦 PCA model initialized (sklearnex version).")
 
     def fit(self, X, y=None):
         """
-        Fit the IncrementalPCA model.
+        Fit the PCA model.
 
         Args:
             X (array-like): Training data.
@@ -71,7 +61,7 @@ class IncrementalPCA:
         print(f"📊 Transformed shape: {X_trans.shape}")
         return X_trans
 
-    def save_model(self, path="ipca_model.joblib"):
+    def save_model(self, path="pca_model.joblib"):
         """
         Save the trained model to a file.
 
@@ -81,7 +71,7 @@ class IncrementalPCA:
         joblib.dump(self.model, path)
         print(f"💾 Model saved to {path}")
 
-    def load_model(self, path="ipca_model.joblib"):
+    def load_model(self, path="pca_model.joblib"):
         """
         Load a model from a file.
 
@@ -105,13 +95,13 @@ class IncrementalPCA:
             raise ValueError("❌ Sparse matrix input is not supported for ONNX/IR conversion.")
         print("✅ Model and input are supported for ONNX/IR conversion.")
 
-    def convert_to_ir(self, X_train, model_name="ipca_model"):
+    def convert_to_ir(self, X_train, model_name="pca_model"):
         """
-        Export IncrementalPCA to OpenVINO IR (not implemented).
+        Export PCA to OpenVINO IR (not implemented).
 
         Args:
             X_train (array-like): Training data for shape reference.
             model_name (str): Base name for the exported model files.
         """
         self._check_export_support(X_train)
-        print("⚠️ IncrementalPCA OpenVINO IR export is not implemented yet.")
+        print("⚠️ PCA OpenVINO IR export is not implemented yet.")
