+++ src/otx/backend/openvino/models/sklearn/neighbors/nearest_neighbors.py
@@ -2,10 +2,11 @@
 # SPDX-License-Identifier: Apache-2.0
 """Class definition for the Nearest Neighbors model with OpenVINO optimization for OTX."""
 
-from sklearnex import patch_sklearn
+import os
 import joblib
 from time import time
-from sklearn.neighbors import NearestNeighbors as SkModel
+from sklearnex.neighbors import NearestNeighbors as SkModel
+from sklearn.metrics import pairwise_distances
 
 class NearestNeighbors:
     def __init__(self, *args, use_openvino=True, **kwargs):
@@ -18,18 +19,10 @@ class NearestNeighbors:
             **kwargs: Keyword arguments for sklearn's NearestNeighbors.
         """
         self.use_openvino = use_openvino
-        self._patched = False
-
-        if self.use_openvino:
-            try:
-                patch_sklearn()
-                self._patched = True
-                print("✅ sklearnex patch applied successfully.")
-            except Exception:
-                print("⚠️ sklearnex patch failed.")
+        self._ir_model = None
 
         self.model = SkModel(*args, **kwargs)
-        print("📦 NearestNeighbors model initialized.")
+        print("📦 NearestNeighbors model initialized (sklearnex version).")
 
     def fit(self, X, y=None):
         """
@@ -44,21 +37,53 @@ class NearestNeighbors:
         elapsed = time() - start
         print(f"🚀 Training completed in {elapsed:.4f} seconds.")
 
-    def kneighbors(self, X, n_neighbors=5, return_distance=True):
+    def kneighbors(self, X, n_neighbors=None, return_distance=True):
         """
         Find the K-neighbors of a point.
 
         Args:
-            X (array-like): The query point or points.
-            n_neighbors (int): Number of neighbors to get.
+            X (array-like): Input data.
+            n_neighbors (int, optional): Number of neighbors to get.
             return_distance (bool): Whether to return distances.
 
         Returns:
-            tuple: (distances, indices) if return_distance else indices
+            distances, indices: Arrays representing distances and indices of neighbors.
         """
         return self.model.kneighbors(X, n_neighbors=n_neighbors, return_distance=return_distance)
 
-    def save_model(self, path="nearest_neighbors_model.joblib"):
+    def score(self, X, y=None):
+        """
+        Not supported: NearestNeighbors does not have a score method.
+
+        Args:
+            X (array-like): Input data.
+            y (ignored): Not used.
+
+        Returns:
+            None
+        """
+        print("⚠️ NearestNeighbors does not support scoring.")
+        return None
+
+    def evaluate(self, X, n_neighbors=None):
+        """
+        Evaluate the model by finding neighbors for X.
+
+        Args:
+            X (array-like): Input data.
+            n_neighbors (int, optional): Number of neighbors to get.
+
+        Returns:
+            indices: Indices of neighbors.
+        """
+        start = time()
+        _, indices = self.kneighbors(X, n_neighbors=n_neighbors)
+        elapsed = time() - start
+        print(f"📈 Neighbors found in {elapsed:.4f} seconds.")
+        print(f"🔎 Indices of first 5 queries: {indices[:5]}")
+        return indices
+
+    def save_model(self, path="nearestneighbors_model.joblib"):
         """
         Save the trained model to a file.
 
@@ -68,7 +93,7 @@ class NearestNeighbors:
         joblib.dump(self.model, path)
         print(f"💾 Model saved to {path}")
 
-    def load_model(self, path="nearest_neighbors_model.joblib"):
+    def load_model(self, path="nearestneighbors_model.joblib"):
         """
         Load a model from a file.
 
@@ -78,7 +103,7 @@ class NearestNeighbors:
         self.model = joblib.load(path)
         print(f"📂 Model loaded from {path}")
 
-    def convert_to_ir(self, X_train, model_name="nearest_neighbors"):
+    def convert_to_ir(self, X_train, model_name="nearestneighbors"):
         """
         Not supported: Exporting NearestNeighbors to IR via neural network is not possible.
